backend/requirements.txt
# FastAPI & Server
fastapi
uvicorn[standard]
python-multipart
python-dotenv
pydantic

# LangChain Core
langchain
langchain-core
langchain-community
langchain-google-genai
langchain-groq

# Vector Store & RAG
faiss-cpu
chromadb

# Document Loaders
pypdf
unstructured
python-docx

# YouTube Transcript
youtube-transcript-api
pytube

# Web Scraping
requests
beautifulsoup4
duckduckgo-search

# Data Analysis
pandas
numpy
matplotlib
seaborn
scipy
scikit-learn

# TensorFlow Lite / DeepFake Detection
tflite-runtime
tensorflow
Pillow
opencv-python-headless

# Embeddings
sentence-transformers
